requirements.txt
finnhub-python
discord
